requirements.txt
pandas>=1.3.0
lockss-pybasic==0.2.0.dev6